Updated the Auto and Added Some other Options.
Updated the Auto to change how to auto will handle the Delivery of the
Gear.

Added an Alliance_ID VI to Store data As a Global.
Added a Counter_Reset VI to handle the Reset of the Counter(s).

Also I added 3 Libs to handle other operations that make it easy to use
the Dashbaord for feed back.